Isolating main entry function from the rest of the code by adding daemon_main.c Cause: Cannot execute units tests if there is two main functions in a single file (includes included!)
